+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Change tar_render_manuscript() to use Typst instead of LaTeX/docx
 
+* Add `ttab()` for creating Typst tables in output
+
 # CMORprojects 0.3.5
 
 * Correct identification of linked Quarto source documents in subfolders